Make plan detail creation inline
